refactor(toolbar): extract `synchronizePlacementOnUpdate` function
