+++ packages/astro/src/runtime/client/dev-toolbar/apps/settings.ts
@@ -1,7 +1,11 @@
 import type { DevToolbarApp } from '../../../../@types/astro.js';
 import { type Settings, settings } from '../settings.js';
-import { type Placement, isValidPlacement, placements } from '../ui-library/window.js';
-import { closeOnOutsideClick, createWindowElement } from './utils/window.js';
+import { isValidPlacement, placements } from '../ui-library/window.js';
+import {
+	closeOnOutsideClick,
+	createWindowElement,
+	synchronizePlacementOnUpdate,
+} from './utils/window.js';
 
 interface SettingRow {
 	name: string;
@@ -71,18 +75,8 @@ export default {
 
 		document.addEventListener('astro:after-swap', createSettingsWindow);
 
-		eventTarget.addEventListener('placement-updated', (evt) => {
-			if (!(evt instanceof CustomEvent)) {
-				return;
-			}
-			const windowElement = canvas.querySelector('astro-dev-toolbar-window');
-			if (!windowElement) {
-				return;
-			}
-			const event: CustomEvent<{ placement: Placement }> = evt;
-			windowElement.placement = event.detail.placement;
-		});
 		closeOnOutsideClick(eventTarget);
+		synchronizePlacementOnUpdate(eventTarget, canvas);
 
 		function createSettingsWindow() {
 			const windowElement = createWindowElement(
