+++ packages/astro/src/runtime/client/dev-toolbar/apps/utils/window.ts
@@ -1,4 +1,5 @@
 import { settings } from '../../settings.js';
+import type { Placement } from '../../ui-library/window.js';
 
 export function createWindowElement(content: string, placement = settings.config.placement) {
 	const windowElement = document.createElement('astro-dev-toolbar-window');
@@ -31,5 +32,19 @@ export function closeOnOutsideClick(
 		} else {
 			document.removeEventListener('click', onPageClick, true);
 		}
+	});
+}
+
+export function synchronizePlacementOnUpdate(eventTarget: EventTarget, canvas: ShadowRoot) {
+	eventTarget.addEventListener('placement-updated', (evt) => {
+		if (!(evt instanceof CustomEvent)) {
+			return;
+		}
+		const windowElement = canvas.querySelector('astro-dev-toolbar-window');
+		if (!windowElement) {
+			return;
+		}
+		const event: CustomEvent<{ placement: Placement }> = evt;
+		windowElement.placement = event.detail.placement;
 	});
 }
